requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Cython==0.29.13
-e git+https://github.com/thtrieu/darkflow.git@b2aee0000cd2a956b9f1de6dbfef94d53158b7d8#egg=darkflow
decorator==4.4.0
Flask==1.1.2
Flask-Cors==3.0.8
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.3
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.3
Pillow==7.1.2
protobuf==3.11.3
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.0.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.4.1
simplejson==3.16.0
six==1.14.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
